+++ app/src/main/java/woowacourse/paint/adapter/ColorAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.adapter
 
 import android.view.ViewGroup
 import androidx.annotation.ArrayRes
