+++ app/src/main/java/woowacourse/paint/adapter/ColorViewHolder.kt
@@ -1,8 +1,9 @@
-package woowacourse.paint
+package woowacourse.paint.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import woowacourse.paint.R
 import woowacourse.paint.databinding.ItemColorBinding
 
 class ColorViewHolder(
