+++ app/src/main/java/woowacourse/paint/model/Painting.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.model
 
 import android.graphics.Canvas
 import android.graphics.Paint
