+++ app/src/main/java/woowacourse/paint/model/Tools.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.model
 
 enum class Tools {
     LINE, RECTANGLE, CIRCLE, ERASER
